Almost complete

+++ Monster.java
@@ -1,6 +1,6 @@
 public interface Monster {
-    public void sethp(double hp);
-    public double gethp();
+    public void sethp(int hp);
+    public int gethp();
     public void monster_attack(Hero h,int p);
     public int getLevel();
 }
